refactor: #76 위키 공통 컴포넌트화
